backend/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
PyJWT[crypto]>=2.7
requests>=2.32
